[05/18]리뷰페이지 연동 완료
