+++ src/pages/Item/OwnerReview.jsx
@@ -1,0 +1,345 @@
+import axios from 'axios';
+import React, { useState, useEffect } from "react";
+import styled from 'styled-components';
+import { createGlobalStyle } from 'styled-components';
+import apiClient from "../../path/apiClient";
+import { useNavigate,useParams,useLocation } from 'react-router-dom';
+
+
+const OwnerReview = () => {
+  const { userId } = useParams();
+  const navigate = useNavigate();
+  const [reviews, setReviews] = useState([]);
+  const [activeReviewType, setActiveReviewType] = useState('rental');
+  const location = useLocation();
+  const { userProfile, userNickName, averageScore } = location.state || {}; // location.state에서 데이터 추출
+
+
+  const [visibleCounts, setVisibleCounts] = useState({
+    rental: 2,
+    deal: 2
+  });
+
+  useEffect(() => {
+    fetchReviews();
+  }, [userId, activeReviewType]);
+
+  const fetchReviews = async () => {
+    const endpoint = activeReviewType === 'rental'
+      ? `/review/reviewDetails/itemOwner/${userId}`
+      : `/review/reviewDetails/renter/${userId}`;
+    
+    try {
+      const { data } = await apiClient.get(endpoint);
+      setReviews(data.content || []);
+    } catch (error) {
+      console.error('Failed to fetch reviews:', error);
+    }
+  };
+
+  const calculateAverageScore = (ownerScore, renterScore) => {
+    let validScores = [];
+    if (ownerScore) validScores.push(ownerScore);
+    if (renterScore) validScores.push(renterScore);
+    return validScores.length ? validScores.reduce((acc, score) => acc + score, 0) / validScores.length : 0;
+  };
+  const StarRating = ({ score }) => {
+    const fullStars = Math.floor(score);
+    const emptyStars = 5 - fullStars;
+  
+    return (
+      <div>
+        {Array.from({ length: fullStars }, (_, index) => (
+          <img key={`full-${index}`} src="/assets/img/YellowStar.png" alt="Full Star" style={{ width: '20px' }} />
+        ))}
+        {Array.from({ length: emptyStars }, (_, index) => (
+          <img key={`empty-${index}`} src="/assets/img/GrayStar.png" alt="Empty Star" style={{ width: '20px' }} />
+        ))}
+      </div>
+    );
+  };
+
+  const handleLoadMore = () => {
+    setVisibleCounts(prevCounts => ({
+      ...prevCounts,
+      [activeReviewType]: prevCounts[activeReviewType] + 3
+    }));
+  };
+
+  const handleGoBack = () => {
+    console.log('돌아가기 버튼 클릭');
+    navigate(-1);
+  };
+
+  return (
+    <>
+      <GlobalStyle />            
+      <BackButton src="/assets/img/BackArrow.png" alt="Back" onClick={handleGoBack} />
+      <Container>
+        <ProfileContainer>
+          <ProfileInfo>
+          <Avatar src={userProfile || "/assets/img/Profile.png"} alt="Profile" />
+            <ProfileDetails>
+              <Nickname>{userNickName}</Nickname>
+              <RatingDetail>
+                <Rating src="/assets/img/Star.png" alt="Star" />
+                <Ratingavg>{averageScore}</Ratingavg>
+              </RatingDetail>
+            </ProfileDetails>
+          </ProfileInfo>
+        </ProfileContainer>
+        <RentingInfoBox>
+          <Reserved onClick={() => setActiveReviewType('rental')} isActive={activeReviewType === 'rental'}>
+            대여 후기
+          </Reserved>
+          <Returned onClick={() => setActiveReviewType('deal')} isActive={activeReviewType === 'deal'}>
+            거래 후기
+          </Returned>
+        </RentingInfoBox>
+        <ReviewList>
+          {reviews.slice(0, visibleCounts[activeReviewType]).map((review, index) => (
+            <ReviewItem key={index}>
+              <ProfileItem>
+                <ProfileImg src={activeReviewType === 'rental' ? review.renterProfile : review.ownerProfile} alt="Profile" />
+                <NickNameAndRating>
+                  <UserNickname>{activeReviewType === 'rental' ? review.renterNickName : review.ownerNickName}</UserNickname>
+                  <StarRating score={review.reviewScore} />
+                </NickNameAndRating>
+              </ProfileItem>
+              <ItemTitle onClick={() => navigate(`/rent/itemdetail/${review.itemId}`)}>
+                {review.itemTitle}
+              </ItemTitle>
+              <Comment>{review.reviewComment}</Comment>
+            </ReviewItem>
+          ))}
+        </ReviewList>
+        {visibleCounts[activeReviewType] < reviews.length && (
+          <MoreViewButton onClick={handleLoadMore}>더 보기</MoreViewButton>
+        )}
+      </Container>
+    </>
+  );
+};
+  export default OwnerReview;
+
+export const GlobalStyle = createGlobalStyle`
+  html, body, #root {
+      height: 100%;
+      margin: 0;
+      padding: 0;
+      display: flex;
+      flex-direction: column;
+      font-family: "Pretendard";
+      background-color: #000; // body 전체의 배경색을 검은색으로 설정
+  }
+
+  /* 스크롤바 전체 스타일 */
+  ::-webkit-scrollbar {
+    width: 0.5rem;
+  }
+
+  /* 스크롤바 트랙(바탕) 스타일 */
+  ::-webkit-scrollbar-track {
+    background: transparent; /* 트랙의 배경색 */
+  }
+
+  /* 스크롤바 핸들(움직이는 부분) 스타일 */
+  ::-webkit-scrollbar-thumb {
+    background: #00FFE0; /* 핸들의 배경색 */
+    border-radius: 5px;
+  }
+`;
+
+const BackButton = styled.img`
+  width: 2rem;
+  height: 2rem;
+  margin-left: 15rem;
+  cursor: pointer;
+  margin-bottom: -4rem;
+`;
+
+const RentingInfoBox = styled.div`
+  background: #1F1F1F;
+  display: flex;
+  justify-content: space-between;
+  width: 59.5rem;
+  height: 4rem;
+  margin-top: 0rem;
+  align-items: center;
+  border-radius: 10px;
+`;
+
+const Reserved = styled.button`
+  flex-grow: 1; /* 자식 요소들의 너비를 동일하게 설정 */
+  height: 100%;
+  background-color: transparent;
+  border: none;
+  font-size: 1.2rem;
+  font-weight: 600;
+  cursor: pointer;
+  color: ${(props) => (props.isActive ? '#00FFE0' : 'white')};
+  border-bottom: ${(props) => (props.isActive ? '2px solid #00FFE0' : 'none')};
+  border-radius: 5px;
+`;
+
+const Returned = styled.button`
+  flex-grow: 1; /* 자식 요소들의 너비를 동일하게 설정 */
+  height: 100%;
+  background-color: transparent;
+  border: none;
+  font-size: 1.2rem;
+  font-weight: 600;
+  cursor: pointer;
+  color: ${(props) => (props.isActive ? '#00FFE0' : 'white')};
+  border-bottom: ${(props) => (props.isActive ? '2px solid #00FFE0' : 'none')};
+  border-radius: 5px;
+`;
+
+const Container = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  font-family: "Pretendard";
+`;
+
+const ProfileContainer = styled.div`
+  width: 58rem;
+  height: 6.5rem;
+  margin-bottom: 15px;
+  border-radius: 5px;
+  padding: 10px;
+  display: flex;
+  margin-top: 4rem;
+  align-items: center;
+  justify-content: center;
+`;
+
+const ProfileInfo = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  margin-bottom: 10px;
+  margin-right: 2rem;
+`;
+
+const Avatar = styled.img`
+  width: 4.6875rem;
+  height: 4.6875rem;
+  border-radius: 50%;
+`;
+
+const ProfileDetails = styled.div`
+  display: flex;
+  flex-direction: column;
+  text-align: center;
+  align-items: center;
+`;
+
+const RatingDetail = styled.div`
+  display: flex;
+  margin-top: 1rem;
+  margin-left: -1rem;
+  margin-bottom: 2rem;
+`;
+
+const Rating = styled.img`
+  width: 1.3rem;
+  height: 1.3rem;
+`;
+
+const Ratingavg = styled.span` 
+  color: #FFF;
+  font-family: "Pretendard";
+  font-size: 1.1rem;
+  font-style: normal;
+  font-weight: 500;
+  margin-left: 1rem;
+`;
+
+const Nickname = styled.span`
+  color: #00FFE0;
+  font-family: "Pretendard";
+  font-size: 1.2rem;
+  font-style: normal;
+  font-weight: 600;
+  margin-top: 1rem;
+`;
+
+const ReviewList = styled.div`
+`;
+
+const ProfileItem = styled.div `
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+`;
+
+const NickNameAndRating = styled.div `
+  display: flex;
+  flex-direction: column;
+`;
+
+const UserNickname = styled.span`
+  color: white;
+  font-family: "Pretendard";
+  font-size: 1rem;
+  font-style: normal;
+  font-weight: 500;
+  margin-bottom: 0.3rem;
+`;
+
+const ReviewItem = styled.div`
+  display: flex;
+  flex-direction: column;
+  background: transparent;
+  margin-top: 3rem;
+  width: 60rem;
+  border: none;
+  border-bottom: 2px solid #5A5A5A; /* 각 리뷰 항목 아래에 구분선 추가 */
+`;
+
+const ProfileImg = styled.img`
+  width: 3.5rem; 
+  height: 3.5rem;
+  border-radius: 50%; 
+  margin-right: 20px;
+`;
+
+const ItemTitle = styled.button`
+  font-size: 1rem;
+  color: white; /* 아이템 제목 색상 */
+  margin-top: 1.5rem;
+  margin-left: 1rem;
+  height: 2rem;
+  border: 1px solid #B2B2B2;
+  background-color: transparent;
+  display: inline-block; /* 내용에 따라 너비가 조정되도록 */
+  white-space: nowrap;
+  max-width: 10rem;
+  cursor: pointer;
+`;
+
+const Comment = styled.div`
+  font-size: 1.2rem;
+  color: #FFF;
+  margin-top: 1.5rem;
+  margin-bottom: 2rem;
+  margin-left: 1rem;
+  white-space: pre-wrap; /* 내용에 줄바꿈 적용 */
+`;
+
+const MoreViewButton = styled.button `
+  color: #9C9C9C;
+  background-color: transparent;
+  border: none;
+  font-family: 'Pretendard';
+  font-size: 1.2rem;
+  font-weight: 500;
+  padding-left: 30px; /* 텍스트 왼쪽에 공간 추가 */
+  background-image: url('/assets/img/Plus.png');
+  background-repeat: no-repeat;
+  background-size: 20px 20px; /* 이미지 크기 조정 */
+  background-position: left center; 
+  cursor: pointer;
+  margin-top: 1rem;
+`;
